+++ ComputerArchitectureProject/safeMalloc.h
@@ -4,9 +4,9 @@
 
 #define safeFree(__ptr)	_safeFree(__ptr, &(__ptr), __FILE__, __FUNCTION__, __LINE__);
 
-void* safeMalloc(size_t size);
+void* safeMalloc(unsigned int size);
 
-void* safeCalloc(size_t count, size_t size);
+void* safeCalloc(unsigned int count, unsigned int size);
 
 void _safeFree(void* ptr, void** pPtr, const char* file, const char* func, int line);
